enhancement: move helpers to separate modules
