+++ packages/plugin-http-errors/http-errors.js
@@ -3,6 +3,13 @@
  * A plugin to automatically capture and report HTTP errors
  */
 
+const extractDomain = require('./lib/extract-domain')
+const headersToObject = require('./lib/headers-to-object')
+const parseQueryParams = require('./lib/parse-query-params')
+const shouldCaptureStatusCode = require('./lib/should-capture-status-code')
+const truncate = require('./lib/truncate')
+const xhrResponseHeadersToObject = require('./lib/xhr-response-headers-to-object')
+
 const DEFAULT_HTTP_ERROR_CODES = [{ min: 400, max: 599 }]
 const DEFAULT_MAX_REQUEST_SIZE = 5000
 
@@ -23,101 +30,7 @@ module.exports = (config = {}, global = window) => {
   } = config
 
   // Normalize httpErrorCodes to an array
-  const normalizedCodes = Array.isArray(httpErrorCodes) ? httpErrorCodes : [httpErrorCodes]
-
-  /**
-   * Check if a status code should be captured
-   * @param {number} statusCode - HTTP status code
-   * @returns {boolean} True if should be captured
-   */
-  const shouldCaptureStatusCode = (statusCode) => {
-    return normalizedCodes.some(code => {
-      if (typeof code === 'number') {
-        return code === statusCode
-      }
-      if (code && typeof code === 'object' && 'min' in code && 'max' in code) {
-        return statusCode >= code.min && statusCode <= code.max
-      }
-      return false
-    })
-  }
-
-  /**
-   * Truncate string to max length
-   * @param {string} str - String to truncate
-   * @param {number} maxLength - Maximum length
-   * @returns {string} Truncated string
-   */
-  const truncate = (str, maxLength) => {
-    if (!str || str.length <= maxLength) return str
-    return str.substring(0, maxLength)
-  }
-
-  /**
-   * Parse query parameters from URL
-   * @param {string} url - URL string
-   * @returns {Object} Parsed query parameters
-   */
-  const parseQueryParams = (url) => {
-    try {
-      const urlObj = new URL(url)
-      const params = {}
-      urlObj.searchParams.forEach((value, key) => {
-        params[key] = value
-      })
-      return params
-    } catch (e) {
-      return {}
-    }
-  }
-
-  const responseHeadersToObject = (headersString) => {
-    if (!headersString) return {}
-    const arr = headersString.trim().split(/[\r\n]+/)
-    const headerMap = {}
-    arr.forEach((line) => {
-      const parts = line.split(': ')
-      const header = parts.shift()
-      const value = parts.join(': ')
-      headerMap[header] = value
-    })
-    return headerMap
-  }
-
-  /**
-   * Convert Headers object to plain object
-   * @param {Headers} headers - Headers object
-   * @returns {Object} Plain object with header key-value pairs
-   */
-  const headersToObject = (headers) => {
-    if (!headers) return {}
-
-    const obj = {}
-    if (headers.entries) {
-      for (const [key, value] of headers.entries()) {
-        obj[key] = value
-      }
-    } else if (headers.forEach) {
-      headers.forEach((value, key) => {
-        obj[key] = value
-      })
-    }
-    return obj
-  }
-
-  /**
-   * Extract domain from URL
-   * @param {string} url - URL string
-   * @returns {string} Domain
-   */
-  const extractDomain = (url) => {
-    try {
-      const urlObj = new URL(url)
-      return urlObj.host
-    } catch (e) {
-      return 'unknown'
-    }
-  }
+  const normalizedStatusCodes = Array.isArray(httpErrorCodes) ? httpErrorCodes : [httpErrorCodes]
 
   let restoreFunctions = []
   const plugin = {
@@ -166,7 +79,7 @@ module.exports = (config = {}, global = window) => {
 
       function handleHttpError (startContext, endContext) {
         // Check if we should capture this status code
-        if (!shouldCaptureStatusCode(endContext.status)) {
+        if (!shouldCaptureStatusCode(normalizedStatusCodes, endContext.status)) {
           return
         }
 
@@ -198,7 +111,7 @@ module.exports = (config = {}, global = window) => {
         if (endContext.response && endContext.response.headers) {
           responseHeaders = headersToObject(endContext.response.headers)
         } else if (endContext.xhr && typeof endContext.xhr.getAllResponseHeaders === 'function') {
-          responseHeaders = responseHeadersToObject(endContext.xhr.getAllResponseHeaders())
+          responseHeaders = xhrResponseHeadersToObject(endContext.xhr.getAllResponseHeaders())
         }
 
         // Create request and response objects for callback
