+++ packages/plugin-http-errors/lib/parse-query-params.js
@@ -1,0 +1,19 @@
+/**
+   * Parse query parameters from URL
+   * @param {string} url - URL string
+   * @returns {Object} Parsed query parameters
+   */
+const parseQueryParams = (url) => {
+  try {
+    const urlObj = new URL(url)
+    const params = {}
+    urlObj.searchParams.forEach((value, key) => {
+      params[key] = value
+    })
+    return params
+  } catch (e) {
+    return {}
+  }
+}
+
+module.exports = parseQueryParams
