+++ packages/plugin-http-errors/lib/should-capture-status-code.js
@@ -1,0 +1,19 @@
+/**
+   * Check if a status code should be captured
+   * @param {number[]} codes - Array of HTTP status codes
+   * @param {number} statusCode - HTTP status code
+   * @returns {boolean} True if should be captured
+   */
+const shouldCaptureStatusCode = (codes, statusCode) => {
+  return codes.some(code => {
+    if (typeof code === 'number') {
+      return code === statusCode
+    }
+    if (code && typeof code === 'object' && 'min' in code && 'max' in code) {
+      return statusCode >= code.min && statusCode <= code.max
+    }
+    return false
+  })
+}
+
+module.exports = shouldCaptureStatusCode
