+++ packages/plugin-http-errors/lib/xhr-response-headers-to-object.js
@@ -1,0 +1,14 @@
+const xhrResponseHeadersToObject = (headersString) => {
+  if (!headersString) return {}
+  const arr = headersString.trim().split(/[\r\n]+/)
+  const headerMap = {}
+  arr.forEach((line) => {
+    const parts = line.split(': ')
+    const header = parts.shift()
+    const value = parts.join(': ')
+    headerMap[header] = value
+  })
+  return headerMap
+}
+
+module.exports = xhrResponseHeadersToObject
